Add debug command to display specific icons
